requirements.txt
fastapi
uvicorn
pydantic
transformers
torch  # Required for Whisper
openai-whisper

